Fix it so that it runs again
