+++ data/xyz.isantop.repoman.policy
@@ -13,7 +13,7 @@
       <allow_inactive>no</allow_inactive>
       <allow_active>auth_admin_keep</allow_active>
     </defaults>
-    <annotate key="org.freedesktop.policykit.exec.path">/usr/local/bin/xyz.isantop.repoman</annotate>
+    <annotate key="org.freedesktop.policykit.exec.path">/usr/bin/xyz.isantop.repoman</annotate>
     <annotate key="org.freedesktop.policykit.exec.allow_gui">TRUE</annotate>
   </action>
