+++ pkexec
@@ -1,2 +1,2 @@
-#!/bin/shì
+#!/bin/sh
 pkexec env DISPLAY=$DISPLAY XAUTHORITY=$XAUTHORITY xyz.isantop.repoman /etc/rc.local
