+++ repoman/window.py
@@ -21,14 +21,9 @@
 import gi
 gi.require_version('Gtk', '3.0')
 from gi.repository import Gtk, Gdk, GLib
-try:
-    from headerbar import Headerbar
-    from stack import Stack
-    from ppa import PPA
-except ImportError:
-    from repoman.headerbar import Headerbar
-    from repoman.stack import Stack
-    from repoman.ppa import PPA
+from .headerbar import Headerbar
+from .stack import Stack
+from .ppa import PPA
 
 class Window(Gtk.Window):
 
@@ -46,198 +41,10 @@ class Window(Gtk.Window):
         self.screen = Gdk.Screen.get_default()
         self.css_provider = Gtk.CssProvider()
         try:
-            self.css_provider.load_from_path('style.css')
+            self.css_provider.load_from_path('data/style.css')
         except GLib.Error:
-            self.css_provider.load_from_path('/usr/local/bin/repoman/style.css')
+            self.css_provider.load_from_path('/usr/share/repoman/style.css')
         self.context = Gtk.StyleContext()
         self.context.add_provider_for_screen(self.screen, self.css_provider,
           Gtk.STYLE_PROVIDER_PRIORITY_USER)
 
-class DeleteDialog(Gtk.Dialog):
-
-    ppa_name = False
-
-    def __init__(self, parent):
-        Gtk.Dialog.__init__(self, "Remove Source", parent, 0,
-                            (Gtk.STOCK_CANCEL, Gtk.ResponseType.CANCEL,
-                             Gtk.STOCK_REMOVE, Gtk.ResponseType.OK),
-                             modal=1, use_header_bar=1)
-
-        content_area = self.get_content_area()
-
-        content_grid = Gtk.Grid()
-        content_grid.set_margin_top(24)
-        content_grid.set_margin_left(24)
-        content_grid.set_margin_right(24)
-        content_grid.set_margin_bottom(24)
-        content_grid.set_column_spacing(36)
-        content_grid.set_row_spacing(12)
-        content_area.add(content_grid)
-
-        delete_image = Gtk.Image.new_from_icon_name("dialog-warning-symbolic",
-                                                Gtk.IconSize.DIALOG)
-        content_grid.attach(delete_image, 0, 0, 1, 2)
-
-        delete_label = Gtk.Label("Are you sure you want to remove this source?")
-        Gtk.StyleContext.add_class(delete_label.get_style_context(), "h2")
-        content_grid.attach(delete_label, 1, 0, 1, 1)
-
-        delete_explain = Gtk.Label("If you remove this source, you will need " +
-                                   "to add it again to continue using it.\n" +
-                                   "Any software you've installed from this " +
-                                   "source will remain installed.")
-        content_grid.attach(delete_explain, 1, 1, 1, 1)
-
-        Gtk.StyleContext.add_class(self.get_widget_for_response(Gtk.ResponseType.OK).get_style_context(),
-                                   "destructive-action")
-
-        self.show_all()
-
-class AddDialog(Gtk.Dialog):
-
-    ppa_name = False
-
-    def __init__(self, parent):
-        Gtk.Dialog.__init__(self, "Add Source", parent, 0,
-                            (Gtk.STOCK_CANCEL, Gtk.ResponseType.CANCEL,
-                             Gtk.STOCK_ADD, Gtk.ResponseType.OK),
-                             modal=1, use_header_bar=1)
-
-        content_area = self.get_content_area()
-
-        content_grid = Gtk.Grid()
-        content_grid.set_margin_top(24)
-        content_grid.set_margin_left(12)
-        content_grid.set_margin_right(12)
-        content_grid.set_margin_bottom(12)
-        content_grid.set_row_spacing(6)
-        content_grid.set_halign(Gtk.Align.CENTER)
-        content_grid.set_hexpand(True)
-        content_area.add(content_grid)
-
-        add_title = Gtk.Label("Enter Source Line")
-        Gtk.StyleContext.add_class(add_title.get_style_context(), "h2")
-        content_grid.attach(add_title, 0, 0, 1, 1)
-
-        add_label = Gtk.Label("e.g. ppa:mirkobrombin/ppa")
-        content_grid.attach(add_label, 0, 1, 1, 1)
-
-        self.ppa_entry = Gtk.Entry()
-        self.ppa_entry.set_placeholder_text("Source Line")
-        self.ppa_entry.set_activates_default(True)
-        self.ppa_entry.set_width_chars(50)
-        self.ppa_entry.set_margin_top(12)
-        content_grid.attach(self.ppa_entry, 0, 2, 1, 1)
-
-        Gtk.StyleContext.add_class(self.get_widget_for_response(Gtk.ResponseType.OK).get_style_context(),
-                                   "suggested-action")
-
-        self.show_all()
-
-
-class EditDialog(Gtk.Dialog):
-
-    ppa_name = False
-
-    def __init__(self,
-                 parent,
-                 repo_disabled,
-                 repo_type,
-                 repo_uri,
-                 repo_version,
-                 repo_component,
-                 repo_archs):
-        Gtk.Dialog.__init__(self, "Modify Source", parent, 0,
-                            (Gtk.STOCK_CANCEL, Gtk.ResponseType.CANCEL,
-                             Gtk.STOCK_SAVE, Gtk.ResponseType.OK),
-                             modal=1, use_header_bar=1)
-
-        self.ppa = PPA(self)
-        self.parent = parent
-
-        content_area = self.get_content_area()
-
-        content_grid = Gtk.Grid()
-        content_grid.set_margin_top(24)
-        content_grid.set_margin_left(12)
-        content_grid.set_margin_right(12)
-        content_grid.set_margin_bottom(12)
-        content_grid.set_column_spacing(12)
-        content_grid.set_row_spacing(6)
-        content_grid.set_halign(Gtk.Align.FILL)
-        content_grid.set_hexpand(True)
-        content_area.add(content_grid)
-
-        type_label = Gtk.Label("Type")
-        type_label.set_halign(Gtk.Align.END)
-        uri_label = Gtk.Label("URI")
-        uri_label.set_halign(Gtk.Align.END)
-        version_label = Gtk.Label("Version")
-        version_label.set_halign(Gtk.Align.END)
-        component_label = Gtk.Label("Component")
-        component_label.set_halign(Gtk.Align.END)
-        enabled_label = Gtk.Label("Enabled")
-        enabled_label.set_halign(Gtk.Align.END)
-        content_grid.attach(type_label, 0, 0, 1, 1)
-        content_grid.attach(uri_label, 0, 1, 1, 1)
-        content_grid.attach(version_label, 0, 2, 1, 1)
-        content_grid.attach(component_label, 0, 3, 1, 1)
-        content_grid.attach(enabled_label, 0, 4, 1, 1)
-
-        self.type_box = Gtk.ComboBoxText()
-        self.type_box.append("deb", "Binary")
-        self.type_box.append("deb-src", "Source code")
-        self.type_box.set_active_id(repo_type)
-        content_grid.attach(self.type_box, 1, 0, 1, 1)
-
-        self.uri_entry = Gtk.Entry()
-        self.uri_entry.set_hexpand(True)
-        self.uri_entry.set_placeholder_text("https://ppa.launchpad.net/...")
-        self.uri_entry.set_text(repo_uri)
-        self.uri_entry.set_activates_default(False)
-        self.uri_entry.set_width_chars(50)
-        content_grid.attach(self.uri_entry, 1, 1, 1, 1)
-
-        self.version_entry = Gtk.Entry()
-        self.version_entry.set_placeholder_text("artful")
-        self.version_entry.set_text(repo_version)
-        self.version_entry.set_activates_default(False)
-        content_grid.attach(self.version_entry, 1, 2, 1, 1)
-
-        self.component_entry = Gtk.Entry()
-        self.component_entry.set_placeholder_text("main")
-        self.component_entry.set_text(repo_component[0])
-        self.component_entry.set_activates_default(False)
-        content_grid.attach(self.component_entry, 1, 3, 1, 1)
-
-        self.enabled_switch = Gtk.Switch()
-        self.enabled_switch.set_halign(Gtk.Align.START)
-        self.enabled_switch.set_active(not repo_disabled)
-        content_grid.attach(self.enabled_switch, 1, 4, 1, 1)
-
-
-        remove_button = Gtk.Button.new_with_label("Remove Source")
-        Gtk.StyleContext.add_class(remove_button.get_style_context(),
-                                   "destructive-action")
-        remove_button.set_margin_top(12)
-        remove_button.connect("clicked", self.on_remove_button_clicked)
-        content_grid.attach(remove_button, 0, 5, 1, 1)
-
-        Gtk.StyleContext.add_class(self.get_widget_for_response(Gtk.ResponseType.OK).get_style_context(),
-                                   "suggested-action")
-
-        self.show_all()
-
-    def on_remove_button_clicked(self, widget):
-        print("Remove Clicked")
-        dialog = DeleteDialog(self)
-        response = dialog.run()
-
-        if response == Gtk.ResponseType.OK:
-            self.ppa.remove(self.parent.stack.list_all.ppa_name)
-            dialog.destroy()
-            self.destroy()
-        else:
-            dialog.destroy()
-
-
