+++ setup.py
@@ -7,6 +7,7 @@ install_data = [('share/metainfo', ['data/xyz.isantop.repoman.appdata.xml']),
                 ('share/applications', ['data/xyz.isantop.repoman.desktop']),
                 ('bin/repoman',['data/style.css']),
                 ('bin/repoman',['repoman/constants.py']),
+                ('bin/repoman',['repoman/dialog.py']),
                 ('bin/repoman',['repoman/headerbar.py']),
                 ('bin/repoman',['repoman/list.py']),
                 ('bin/repoman',['repoman/main.py']),
