+++ xyz.isantop.repoman
@@ -19,24 +19,6 @@
 '''
 import os
 import sys
-
-# Get launch script dir
-launch_dir = os.path.dirname(os.path.abspath(sys.argv[0]))
-
-# Update sys.path to include modules
-if launch_dir == "/usr/bin":
-    modules_path = "/usr/share/xyz.isantop.repoman/repoman"
-else:
-    modules_path = launch_dir + "/repoman"
-
-sys.path.insert(0, modules_path)
-"""
-try:
-    from repoman import main
-except ImportError:
-    print("Failed to import module main.py!")
-    print("Installation was assumed to be at:", modules_path)
-    sys.exit(1)"""
 from repoman import main
